py/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ucp-server-py"
version = "1.0.0"
description = "Generate UCP-compliant servers for merchants from a simple schema"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["ucp", "commerce", "checkout", "agentic", "ai-commerce"]
dependencies = [
    "flask>=3.0.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "stripe>=7.0.0",
    "paypalserversdk>=1.0.0",
]

[project.scripts]
ucp-server = "ucp_server.cli:main"

[project.urls]
Homepage = "https://ucp.dev"
Repository = "https://github.com/user/ucp-server"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-flask>=1.3.0",
]
